pyproject.toml
[tool.poetry]
name = "inventix"
version = "0.1.0"
description = ""
authors = ["ZeroTwo-M1000 <BioArtex@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
loguru = "^0.7.2"
pyjwt = "^2.8.0"
prisma = "^0.13.0"
requests = "^2.31.0"
lxml = "^5.1.0"
beautifulsoup4 = "^4.12.3"
fastapi = {extras = ["all"], version = "^0.110.0"}


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
